Pinned down. disk_total_space() and disk_free_space() return float in PHP — always, even on 64-bit systems. With declare(strict_types=1), passing that float into human_bytes(int $bytes) throws a fatal TypeError, crashing the whole endpoint before it can output anything

+++ webui/api/vps_info.php
@@ -33,8 +33,8 @@ $memAvail = $mem['MemAvailable'] ?? 0;
 $memUsed = $memTotal - $memAvail;
 
 // --- Disk (root fs) ---
-$diskTotal = @disk_total_space('/') ?: 0;
-$diskFree  = @disk_free_space('/') ?: 0;
+$diskTotal = (int) (@disk_total_space('/') ?: 0);
+$diskFree  = (int) (@disk_free_space('/') ?: 0);
 $diskUsed  = $diskTotal - $diskFree;
 
 // --- Uptime ---
